Backend/requirements.txt
Django==5.0.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
python-dotenv==1.0.0
mysqlclient==2.2.7
Pillow==11.2.1
requests==2.31.0
